+++ Map/Controller/MainMenuViewController.m
@@ -13,6 +13,12 @@
 #import "RQHero.h"
 #import "CreateHeroViewController.h"
 
+// controllers
+#import "TrekListViewController.h"
+#import "WeightLogViewController.h"
+#import "SettingsViewController.h"
+#import "DeveloperToolboxViewController.h"
+
 @implementation MainMenuViewController
 
 @synthesize delegate;
@@ -23,7 +29,7 @@
 - (id)init
 {
 	if (self = [super initWithNibName:@"MainMenuView" bundle:nil]) {
-		
+		self.wantsFullScreenLayout = YES;
 	}
 	return self;
 }
@@ -45,7 +51,7 @@
 
 - (void)viewWillAppear:(BOOL)animated
 {
-	[[UIApplication sharedApplication] setStatusBarHidden:NO withAnimation:UIStatusBarAnimationFade];
+	[[UIApplication sharedApplication] setStatusBarHidden:YES withAnimation:UIStatusBarAnimationFade];
 	[super viewWillAppear:animated];
 }
 
@@ -73,17 +79,57 @@
 
 - (IBAction)optionsButtonPressed:(id)sender
 {
-	[delegate mainMenuViewControllerOptionsButtonPressed:self];
+	SettingsViewController *settingsVC = [[SettingsViewController alloc] init];
+	[settingsVC setDelegate:self];
+	[self presentModalViewController:settingsVC animated:YES];
+	[settingsVC release];
 }
 
 - (IBAction)logButtonPressed:(id)sender
 {
-	[delegate mainMenuViewControllerTreksButtonPressed:self];
+	//[delegate mainMenuViewControllerTreksButtonPressed:self];
+	// BUILD Trek Log
+	TrekListViewController *trekListVC = [[TrekListViewController alloc] init];
+	NSFetchRequest *request = [[NSFetchRequest alloc] init];
+	NSManagedObjectContext *moc = [[[RQModelController defaultModelController] simpleCoreData] managedObjectContext];
+	[request setEntity:[NSEntityDescription entityForName:@"Trek" inManagedObjectContext:moc]];
+	[request setSortDescriptors:[NSArray arrayWithObjects:[NSSortDescriptor sortDescriptorWithKey:@"date" ascending:NO], nil]];
+	NSFetchedResultsController *resultsController = [[NSFetchedResultsController alloc] initWithFetchRequest:request managedObjectContext:moc sectionNameKeyPath:nil cacheName:@"Treks"];
+	[request release];
+	trekListVC.fetchedResultsController = resultsController;
+	[resultsController release];
+	UINavigationController *trekNavController = [[UINavigationController alloc] initWithRootViewController:trekListVC];
+	[trekListVC release];
+	UIBarButtonItem *doneItem = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemDone target:self action:@selector(doneBrowsingLogBook:)];
+	trekNavController.navigationBar.topItem.rightBarButtonItem = doneItem;
+	
+	// BUILD Weight Log
+	WeightLogViewController *weightVC = [[WeightLogViewController alloc] init];
+	UINavigationController *weightNavController = [[UINavigationController alloc] initWithRootViewController:weightVC];
+	weightNavController.navigationBar.topItem.rightBarButtonItem = doneItem;
+	[weightVC release];
+	
+	[doneItem release];
+	
+	// BUILD tab controller
+	UITabBarController *tabBarController = [[UITabBarController alloc] init];
+	tabBarController.viewControllers = [NSArray arrayWithObjects:trekNavController, weightNavController, nil];
+	[trekNavController release];
+	[weightNavController release];
+	
+	[self presentModalViewController:tabBarController animated:YES];
+	[tabBarController release];
 }
 
 - (IBAction)developerToolboxButtonPressed:(id)sender
 {
-	[delegate mainMenuViewControllerDeveloperToolboxButtonPressed:self];
+	//[delegate mainMenuViewControllerDeveloperToolboxButtonPressed:self];
+	DeveloperToolboxViewController *devVC = [[DeveloperToolboxViewController alloc] init];
+	[devVC setDelegate:self];
+	UINavigationController *developerToolboxNavigationController = [[UINavigationController alloc] initWithRootViewController:devVC];
+	[devVC release];
+	[self presentModalViewController:developerToolboxNavigationController animated:YES];
+	[developerToolboxNavigationController release];
 }
 
 
@@ -96,6 +142,32 @@
 	}
 	
 	[delegate mainMenuViewControllerStoryButtonPressed:self];
+}
+
+- (IBAction)doneBrowsingLogBook:(id)sender
+{	
+	[self dismissModalViewControllerAnimated:YES];
+}
+
+- (IBAction)doneBrowsingDevToolbox:(id)sender
+{	
+	[self dismissModalViewControllerAnimated:YES];
+}
+
+#pragma mark -
+#pragma mark SettingsViewControllerDelegate methods
+
+- (void)settingsViewControllerDidEnd:(SettingsViewController *)controller;
+{
+	[self dismissModalViewControllerAnimated:YES];
+}
+
+#pragma mark -
+#pragma mark DeveloperToolboxViewControllerDelegate Methods
+
+- (void)developerToolboxViewControllerDidEnd:(DeveloperToolboxViewController *)controller
+{
+	[self dismissModalViewControllerAnimated:YES];
 }
 
 @end